+++ tasks.csv
@@ -1,2 +1,2 @@
 Title,Description,Due Date,Due Time,Priority,Category,Completed
-Task1,Sample task,2024-01-12,12:20,low,category 1,False
+task 1,desc,2024-10-02,12:20,low,cat1,False
